bot/requirements.txt
disnake>=2.4.0
PyYAML>=6.0
vgamepad>=0.0.8